Fix some compatibility problems with Postgres
When running against Postgres, the default value for an EnumChoiceField
would be assumed to be the empty string '', which caused problems when
serializing it to the database. Now the default value is assumed to be
None, and migrations against a Postgres DB work again. Additionally,
stricter checking is performed when serializing values for the database,
and better errors are thrown when problems are encountered.